LRCI-3495 Rename package from builds to build
