+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/task/DefaultTask.java
@@ -14,9 +14,9 @@
 
 package com.liferay.jethr0.task;
 
-import com.liferay.jethr0.builds.Build;
-
 import org.json.JSONObject;
+
+import com.liferay.jethr0.build.Build;
 
 /**
  * @author Michael Hashimoto
