+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/task/Task.java
@@ -14,7 +14,7 @@
 
 package com.liferay.jethr0.task;
 
-import com.liferay.jethr0.builds.Build;
+import com.liferay.jethr0.build.Build;
 import com.liferay.jethr0.task.run.TaskRun;
 
 import java.util.List;
