+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/task/TaskFactory.java
@@ -14,12 +14,12 @@
 
 package com.liferay.jethr0.task;
 
-import com.liferay.jethr0.builds.Build;
-
 import java.util.HashMap;
 import java.util.Map;
 
 import org.json.JSONObject;
+
+import com.liferay.jethr0.build.Build;
 
 /**
  * @author Michael Hashimoto
